use /api/v1/statuses/ for status lookup instead of search
Pleroma search is terrible.
Even with resolve=true and passing a direct URL,
it will still attempt a fulltext search in addition to a URL fetch.
On most instances, search is so slow that simply opening a thread in Sengi would take several seconds.
As the search API is intended for "cold" lookups only, we switch to the status lookup endpoint.

This works on both Pleroma and Mastodon.